+++ Emu5/RVEmulator.cs
@@ -5,11 +5,11 @@ namespace Emu5
 {
     class RVEmulator
     {
-        Dictionary<UInt64, UInt64> m_memoryMap; // mapped on 8 bytes to reduce memory consumption
+        Dictionary<UInt32, UInt64> m_memoryMap; // mapped on 8 bytes to reduce memory consumption
 
         public RVEmulator()
         {
-            m_memoryMap = new Dictionary<UInt64, UInt64>();
+            m_memoryMap = new Dictionary<UInt32, UInt64>();
         }
 
         public void Assemble(String code)
